Enum github::activity::IssueEventType

pub enum IssueEventType {
    Closed,
    Reopened,
    Subscribed,
    Merged,
    Referenced,
    Mentioned,
    Assigned,
    Unassigned,
    Labeled,
    Unlabeled,
    Milestoned,
    Demilestoned,
    Renamed,
    Locked,
    Unlocked,
    HeadRefDeleted,
    HeadRefRestored,
    Unknown(String),
}

IssueEventType is an enumuration of all the different types of Issue Events.

Variants

Closed

The issue was closed by the Actor.

Reopened

The issue was reopened by the Actor.

Subscribed

The Actor subscribed to receive notifications for an issue.

Merged

The issue was merged by the Actor.

Referenced

The issue was referenced from a commit message.

Mentioned

The Actor was @mentioned in an issue body.

Assigned

The issue was assigned to the Actor.

Unassigned

The issue was unassigned to the Actor.

Labeled

A label was added to the issue.

Unlabeled

A label was removed from the issue.

Milestoned

The issue was added to a milestone.

Demilestoned

The issue was removed from a milestone.

Renamed

The issue title was changed.

Locked

The issue was locked by the Actor.

Unlocked

The issue was unlocked by the Actor.

HeadRefDeleted

The pull request’s branch was deleted.

HeadRefRestored

The pull request’s branch was restored.

Unknown

Unknown(String) is used as a last resort when an event is unknown. This should never happen, please report/resolve the issue when it does happen.