Module github::error

Structs

ErrorContext
When a negative status was given as a response to a request, there might be one or several error descriptions embedded in the body to tell more about the details of what was wrong. ErrorContext is the representation for each of the errors that are given.

InternalError
InternalError will be given in the form of Result in case something went wrong within this Client Library. It replaces panics so that you can freely choose the behaviour. Please file an issue and/or resolve the bug yourself when you get this error.

RequestError
RequestError will be returned as a Result<T, ClientError> in case a request responds negatively populated by information from both the header and body.

Enums

ClientError
ClientError enumerates all the possible errors that a public client (request) function of this library might be given.

ErrorCode
Documentation References: https://developer.github.com/v3/#client-errors ErrorCode represents the type of error that was reported as a response on a request to th Github API.

ErrorStatus
ErrorStatus represents the status code given in the header of a negative response. Look at const definitions such as STATUS_OK for more information for each value.

Functions

check_status_code
Simplistic function internally used to check if a returned status code is positive. Which means that the request was succesful.