Enum github::error::ErrorStatus

pub enum ErrorStatus {
    BadRequest,
    UnprocessableEntity,
    Forbidden,
    NotFound,
    Unknown(u32),
}

ErrorStatus represents the status code given in the header of a negative response. Look at const definitions such as STATUS_OK for more information for each value.

Variants

BadRequest

UnprocessableEntity

Forbidden

NotFound

Unknown

Methods

impl ErrorStatus

fn new(code: u32) -> ErrorStatus

Simple way to construct an ErrorStatus based on its constant value as defined by the official docs.