Function github::http::get

pub fn get<R: Decodable>(user: &str, url: &str, opts: Option<Vec<(&str, &str)>>) -> Result<(Vec<R>, Response), ClientError>

A simplistic function that wraps around the behaviour of an http get-request as defined in curl. As the library gets more complete, a more complete and complex approach might be needed.