Module github::response

Structs

Page
Page represents a link related to the response

Rate
Rate represents the X-Rate-Limit data provided by the Github v3 API and provided for each response

Response
TODO: implement functionality Response represents the exposed data given with each request and populated by the Github v3 API

Traits

Populatable
Populatable provides functionality to construct an object based on info found in the http response header

Type Definitions

HttpHeaderType
HttpHeaderType defines the type used for raw http headers