Struct github::response::Response

pub struct Response {
    pub resp: HttpHeaderType,
    pub next: Option<Page>,
    pub last: Option<Page>,
    pub first: Option<Page>,
    pub prev: Option<Page>,
    pub rate: Rate,
}

TODO: implement functionality Response represents the exposed data given with each request and populated by the Github v3 API

Fields

resp

the raw response header

next

the immediate next page of result

last

the last page of results

first

the first page of results

prev

the immediate previous page of results

rate

the latest X-Rate-Limit info