Trait github::response::Populatable

pub trait Populatable {
    fn populate(raw_header: &HttpHeaderType) -> Self;
}

Populatable provides functionality to construct an object based on info found in the http response header

Required Methods

fn populate(raw_header: &HttpHeaderType) -> Self

populate constructs a new object based on the information found in the http response header